requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
dnspython==2.8.0
idna==3.11
kafka-python==2.3.1
lxml==6.0.4
numpy==2.4.4
packaging==26.0
pandas==3.0.2
psycopg2-binary==2.9.11
pymongo==4.16.0
python-dateutil==2.9.0.post0
pytrends==4.9.2
requests==2.33.1
six==1.17.0
urllib3==2.6.3
